api_service/pyproject.toml
[project]
name = "api-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pydantic>=2.11.1",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "ruff>=0.11.4",
]
